requirements.txt
jinja2
dotenv
pyyaml
pytest